+++ src/main/kotlin/dev/noctud/plugin/ext/StringExt.kt
@@ -1,5 +1,5 @@
 package dev.noctud.plugin.ext
 
 fun log(message: String) {
-    com.intellij.openapi.diagnostic.Logger.getInstance("NoctudPlugin").warn(message)
+    com.intellij.openapi.diagnostic.Logger.getInstance("NoctudPlugin").debug(message)
 }