+++ CHANGELOG.md
@@ -1,1 +1,10 @@
 # Changelog
+
+## [1.0.0] - 2026-02-13
+
+### Added
+
+- Read-only inspection
+- Support for closure parameter types
+- Map view method autocomplete on Map
+